Make replacement of entities case insensitive.
Some entities would not be replaced with their link wrapped
equivalents since they appeared in different case in the tweet.
For example, a username "Foo" might appear as "@foo" in a tweet,
and therefore would not be replaced since the find value works
off of the canonical screen_name "Foo" of the user.