Use EE2.8's new Cache class, when available
ExpressionEngine added a new Cache[1] class in EE2.8 that is specifically
for persistent caches. ExpressionEngine would like modules to move toward
this, presumably so it can give users more control of cached items.

For this module, that means putting the result of the Twitter API call in
the cache instead of writing to a file.

For EE older than 2.8.0, we still cache to a file.

[1]: http://ellislab.com/expressionengine/user-guide/development/usage/cache.html