Don't output target='' when no target is set
Only output the target attribute for links when it is specified by the
target parameter on the tag.

The ability to set a target on each link was added in 2bf4572. However,
when no target is set, EE_Twitter was still outputting target='', which is
completely useless, and causes warnings and errors in some validators.

Fixes #33.