Add a way to get all tweets from a single list.
The best way to get tweets from many Twitter accounts into a single API
call is to set up a list for those users. So, allow the module to make a
call to get a list's tweets.

There are a few advantages of using this over multiple screen names.

* Only one API call is made.
* It pushes the maintenance burden of managing which Twitter accounts to
  display from the developer to the person in charge of the list. This
  could make it possible for a non-technical adminstrator to add or remove
  users from the list without having to involve a website developer.

Fixes #34.